edit the return type of the sampling interval

+++ msl/equipment/resources/picotech/picoscope/picoscope.py
@@ -442,7 +442,7 @@ class PicoScope(ConnectionSDK):
 
         Returns
         -------
-        :obj:`int`
+        :obj:`float`
             The sampling interval, i.e. dt.
         :obj:`int`
             The number of samples that will be acquired.
